feat : 상품 전체 조회 api를 수정했어요.

+++ src/main/kotlin/org/teamalilm/alilmbe/domain/product/entity/Product.kt
@@ -37,7 +37,7 @@ class Product(
     val price: Int,
 
     @Column(name = "waiting_count")
-    var waitingCount: Long = 0,
+    var waitingCount: Int = 0,
 
     @Embedded
     val productInfo: ProductInfo,
